# fix extra few shots bug